added a star so agent moves to gold and back to start position.
needs better a star optimisation.
a star can also get stuck if it knows the location of the gold but can't find a path to it. if can't find path, agent needs to move to unvisited areas randomly.
random movement of agent can be better, use visited[i][j] to move to spaces not visited